owner hotel ver 4.0 : bổ xung tính năng chọn loại phòng cho các phòng, load danh xách theo loại phòng, thêm loại phòng

+++ app/src/main/java/com/tdc/nhom6/roomio/models/Room.kt
@@ -5,7 +5,7 @@ import com.tdc.nhom6.roomio.R
 
 
 data class Room(
-    val floor: Long = 0,
+    val floor: Int = 0,
     val room_id: String = "",
     val room_number: String = "",
     val room_type_id: String = "",
@@ -14,13 +14,12 @@ data class Room(
     val displayCode: String
         get() = "F${floor}-${room_number}"
 
-    // Hàm lấy ID Drawable cho màu nền dựa trên trạng thái
     fun getBackgroundColorResId(): Int {
         return when (status_id) {
-            "room_occupied" -> R.drawable.bg_chip_red     // Occupied
-            "room_available" -> R.drawable.bg_chip_gray  // Available
-            "room_pending" -> R.drawable.bg_chip_yellow  // Pending
-            "room_fixed" -> R.drawable.bg_chip_blue      // Fixed
+            "room_occupied" -> R.drawable.bg_chip_red
+            "room_available" -> R.drawable.bg_chip_gray
+            "room_pending" -> R.drawable.bg_chip_yellow
+            "room_fixed" -> R.drawable.bg_chip_blue
             else -> R.drawable.bg_chip_gray
         }
     }
